TFT版でGPEEK、GINPの対応、BIN$不具合対応
TFT版でのGPEEK、GINP対応はArduino STM32最新版利用時のみ

fix #24
fix #27

+++ libraries/TTBAS_LIB/tGraphicDev.h
@@ -2,6 +2,7 @@
 // 豊四季Tiny BASIC for Arduino STM32 グラフィック描画デバイス
 // 2017/07/19 by たま吉さん
 // 修正日 2017/08/19, 抽象化に移行、実体はハード依存の各スクリーンクラスにて実装
+// 修正日 2018/08/31, gpeek(),ginp()の戻り値、引数の型を変更(TFT対応のため）
 //
 
 #ifndef __tGraphicDev_h__
@@ -30,8 +31,8 @@ class tGraphicDev {
     virtual void bitmap(int16_t x, int16_t y, uint8_t* adr, uint16_t index, uint16_t w, uint16_t h, uint16_t d, uint8_t rgb=0)  = 0;
     virtual void gscroll(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t mode)  = 0;
     virtual void cscroll(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t d) = 0; 
-    virtual int16_t gpeek(int16_t x, int16_t y)  = 0;
-    virtual int16_t ginp(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t c)  = 0;
+    virtual uint16_t gpeek(int16_t x, int16_t y)  = 0;
+    virtual int16_t ginp(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t c)  = 0;
     virtual void set_gcursor(uint16_t, uint16_t)  = 0;
     virtual void gputch(uint8_t c)  = 0;
